# gate/static/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Restaurant API Gateway",
    "description": "API Gateway для ресторанной системы",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Введите \"Bearer\" за которым следует пробел и JWT токен."
    }
  },
  "paths": {
    "/auth/client/login/init": {
      "post": {
        "tags": ["Auth"],
        "summary": "Инициализация входа клиента",
        "description": "Отправляет код подтверждения на телефон клиента",
        "operationId": "LoginClientInit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "description": "Номер телефона клиента"
                  }
                },
                "required": ["phone"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная инициализация",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/client/login/confirm": {
      "post": {
        "tags": ["Auth"],
        "summary": "Подтверждение входа клиента",
        "description": "Подтверждает вход клиента с помощью кода",
        "operationId": "LoginClientConfirm",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "description": "Номер телефона клиента"
                  },
                  "code": {
                    "type": "string",
                    "description": "Код подтверждения"
                  }
                },
                "required": ["phone", "code"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "integer"
                    },
                    "refresh_token": {
                      "type": "string"
                    },
                    "refresh_token_expires_in": {
                      "type": "integer"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/staff/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Вход сотрудника",
        "description": "Вход сотрудника с помощью email и пароля",
        "operationId": "LoginStaff",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "staff": {
                    "$ref": "#/components/schemas/Staff"
                  },
                  "password": {
                    "type": "string",
                    "description": "Пароль сотрудника"
                  }
                },
                "required": ["staff", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "integer"
                    },
                    "refresh_token": {
                      "type": "string"
                    },
                    "refresh_token_expires_in": {
                      "type": "integer"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/staff/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Регистрация сотрудника",
        "description": "Регистрирует нового сотрудника",
        "operationId": "RegisterStaff",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "staff": {
                    "$ref": "#/components/schemas/Staff"
                  }
                },
                "required": ["staff"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная регистрация",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "staff": {
                      "$ref": "#/components/schemas/Staff"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/validate": {
      "post": {
        "tags": ["Auth"],
        "summary": "Валидация токена",
        "description": "Проверяет валидность JWT токена",
        "operationId": "Validate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "JWT токен для валидации"
                  }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Результат валидации",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Обновление токена",
        "description": "Обновляет access token с помощью refresh token",
        "operationId": "Refresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string",
                    "description": "Refresh token"
                  }
                },
                "required": ["refresh_token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Новые токены",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "integer"
                    },
                    "refresh_token": {
                      "type": "string"
                    },
                    "refresh_token_expires_in": {
                      "type": "integer"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "staff": {
            "$ref": "#/components/schemas/Staff"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ROLE_UNSPECIFIED", "ROLE_CLIENT", "ROLE_ADMIN", "ROLE_STAFF"]
            }
          }
        }
      },
      "Client": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "Staff": {
        "type": "object",
        "properties": {
          "work_email": {
            "type": "string"
          },
          "position": {
            "type": "string"
          }
        }
      }
    }
  }
} 